[4.0] Required fields [a11y]
Asterisks are helpful cues for someone looking at a webpage, but aren’t as helpful when browsing via screen reader. `aria-hidden="true"`should be used to hide an asterisk from a screen reader.

We are only using this for fields rendered with the form layout and this PR adds it to the three plugins with a required field that do not use that layout.