Add visual indicator for failing scheduled tasks
Fixes joomla#36459. Adds a visual indicator to failing tasks in the form
of a warning/danger icon colored yellow. Other color schemes were tested
but this seemed like the best option to me. Open to change.
The visual indicator includes a tooltip for accessibility (not sure if
it needs something else).