[4.1] Modal Buttons :focus [a11y] (#36880)
When you are navigating the admin using the keyboard all the buttons should indicate when they have focus. For toolbar buttons this has been done correctly and the display of the button changes exactly the same on focus and on hover. This is the bootstrap default behaviour.

Unfortunately the buttons in the modals do not use the bootstrap defaults (_I don't know why_) and there is no styling for on focus.

This PR resolves that so the buttons in modals are styled correctly on focus. Displaying a visual indicator of the current focus is a fundamental accessibility issue.

This PR changes the scss so to test either use a prebuilt package or make sure you remember to `npm run build:css`

Open any modal and using the keyboard tab key navigate to the buttons. Before the PR you don't have any indicator when you are on a button. After the PR the button is styled exactly the same as if  you were hovering with your mouse.