[4.1] remove underline on buttons etc
This is a css change so you will need to test with either a pre-built package or by `npm run build:css`

We already remove the underline on links for badges in tables. This extends that to dropdown items and buttons.

To test go to com_users and create a note for a user.

## Before

## After