[4.2] [New Feature] TinyMCE Image Classes
Based on user feedback this PR adds the ability to add "Left, Right & Center" alignment for images using the tinymce image button. This uses the tinymce api and existing tinmyce language strings so should be compatible with all languages and easy to maintain etc.

As a bonus this resolves the current issue that after inserting an image with the media manager when you select the image again you can change the alignment.

This is NOT using inline styles. It uses correct classes but as they are present in cassiopeia but not in the editor.css they have been added to the editor.css to ensure they preview correctly.

To test either use a prebuilt package or apply the pr and rebuild the css `npm run build:css`