[4.2] RFC alias error link
No idea if this is the _best_ way to do this so feedback appreciated. If/when accepted it can be applied to may similar use cases.

pr for #37013

create a menu item called test
Try to create a second menu item also called test

Before

After
And the link opens the existing menu item