[4.1] Fileinfo check
Despite the php extension fileinfo being enabled by default in php we have had quite a few users who are discovering that media manager will not load. This is because they have fileinfo disabled on the server.

This PR adds fileinfo to the list of php settings shown in the system information.

To test
Install the PR and check the php settings page of the system information and you will see the new line "Fileinfo Available" which hopefully has a setting of yes.

o replicate this and therefore to prove the issue and the solution do the following

In your php.ini file locate the line
CODE: SELECT ALL

extension=fileinfo
and comment it out by placing at ; at the beginning of the line.

Then restart apache

Now when you go to media manager you will see the reported error.

Now go back to php.ini and uncomment that line

Restart apache