[4.1] Codemirror Full Screen
This PR re-introduces the on screen instruction how to make codemirror full screen. (No idea why it was removed the pr doesnt help) AND fixes the positioning of the editor when full screen so that the first 5 lines are not hidden.

As there is a css change you will need to use a pre-build package or `npm run build:css`