[4.1] Codemirror search
This PR corrects some obvious path errors in the build script for codemirror and loads the search plugin.

To test this apply the pr and run `npm ci`

Go to the template manager and open a file to edit.

Try ctrl-f to search, ctrl-g to find next, ctrh-h to search replace

Closes #26411