[4.1] Private Messages Emails
PR for #29533 (and more)

Since the component was changed to include the actual message in the email it is no longer a notification so I updated the description show in the mail templates component

Before this PR the subject of the email that was generated was
New Message from <user> at <sitename>

The actual subject of the message was included in the body of the email.
"New Message from <user>" is not needed. We know who the email is from in the usual way and our inbox will show if it is new or not ;)

This also meant that you would have an inbox with multiple messages that all had the exact same title :(

With this PR the subject of the email that is generated is
[<sitename>] - Subject
and the body of the email only contains the body of the message not subject and body.

The messages can of course still be managed and updated with the mail templates component. This is just setting a better default.