[4.1] com_users login div (#37164)
Every view has a single containing div except for the login page which has two divs.

This creates issues when styling if for example you wanted to put a border around the div or a background. Every page but this one will have one box. This one will have two.

To test add the following code immediately after the head in the site template

	```<style>main>[class^="com-"]{background-color:grey; border:5px solid black;}</style>```

### Before


### After