[4.1] jooa11y plugin options (#37196)
* [4.1] jooa11y plugin options

In the initial pull request the removal of some options went too far and stopped the remaining options from working. This resolves that.

This pr also changes the description of the containers to include examples of using classes, divs and roles

Before this PR none of the options were used. Now they are.

* cs

* punct

* CAPS