webp Support for (deprecated) J3 Image Package
A lot of people will continue to use Joomla 3 for some time, yet they want to use WebP images. 
This is a small adjustment to let extensions like media manager accomplish that. Our extensions are using the J4 WebP support for resizing and stuff, and this would be good to offer WebP for all the J3 people still around.