[4.1] Private Messages Emails (#37133)
With this PR the subject of the email that is generated is
[<sitename>] - Subject
and the body of the email only contains the body of the message not subject and body.

The messages can of course still be managed and updated with the mail templates component. This is just setting a better default.