[4.1] content params
In the installation only some of the default params for com_content are setup in the database. It's not really a problem as there are default values BUT it means that where you have a setting in a menu item that is "Use Global" it doesn't display the value.

### For example
#### Before

#### After

This is only for the installation.