[4.1] Add Visual Indicator for Failing Scheduled Tasks (#36765)
* Add visual indicator for failing scheduled tasks

Fixes joomla#36459. Adds a visual indicator to failing tasks in the form
of a warning/danger icon colored yellow. Other color schemes were tested
but this seemed like the best option to me. Open to change.
The visual indicator includes a tooltip for accessibility (not sure if
it needs something else).

* Update indicator-tooltip for consistency

* Add aria-hidden to indicator for a11y

* Fix codestyle

Co-authored-by: Richard Fath <richard67@users.noreply.github.com>

* Fix codestyle (2)

Co-authored-by: Richard Fath <richard67@users.noreply.github.com>

* Apply suggestions from code review

Minor improvements to code quality/contextuality.

Co-authored-by: heelc29 <66922325+heelc29@users.noreply.github.com>

* Apply suggestions from code review

Removes some CSS classes from the failure indicator and improves a language string.

Co-authored-by: Quy <quy@fluxbb.org>

Co-authored-by: heelc29 <66922325+heelc29@users.noreply.github.com>
Co-authored-by: Quy <quy@fluxbb.org>