Change Joomla Update Notification Text
The Current Joomla Update Notification mail is very long and sounds kind of aggressive. Instead of telling the User in Uppercase letters not and do not contact Joomla I changed the text to be more positive and friendly and to contact the site manger. The update mail before repeated itself and was therefore way too long, it confused the non technical admin with some technical words, It sounded very aggressive with uppercase letters and made the update appear scary. I hope next time a Joomla Update is available, the user can get the notification about the "Good News". Thanks!




Hi!
Good news: An updated Joomla! system version for the website {SITENAME} is now available!

Information about the update:
Website name: {SITENAME}
Website address: {URL}
Current Joomla! version: {CURVERSION}
Available Joomla! version: {NEWVERSION}

What do I have to do?
The system update can easily be started via the following link:
{LINK}

To open, please click on the link or copy it into the address bar of your browser.

The system will continue to send reminders until the update is successfully completed.

What has changed?
Detailed information on the latest changes can be found at the link:
{RELEASENEWS}

Why am I receiving this mail?
Joomla! is the software used to operate the website "Website Name". Regular updates are important for the function and security of this website.

Joomla! automatically sends out notifications when a new system version is available.
In order not to receive these notifications in the future, the item "System Joomla! update notification" must be deactivated in the plugin administration of the website.

Who can I contact if I have any questions?
This e-mail is an automatically sent information e-mail from the website {SITENAME}.
If you have any questions about this email or the necessary steps to update, please contact the person who manages your website.