[4.1] Manual Update links underline
This completes the work started in #36134 by making sure that the class is added to ALL the instances of the link and not just the instance on the default page.

Testing is tricky as you will need to force the updater to see a new release so it is probably best to test by code review only.