Deprecate constant JCOMPAT_UNICODE_PROPERTIES
Prepare to remove JCOMPAT_UNICODE_PROPERTIES constants from global namespace and replace it with a static variable within the test function. This is a preparation for PSR12 transformation to satisfy PSR1.Files.SideEffects.FoundWithSymbols rule. Beside that it doesn't make sense to declare a global constant which is only used in the local class.