Refactor the page cache plugin (#36042)
* Refactor the page cache plugin

* Don't try to fix stupid

* Make the plugin class final

* Using low priority

This solves the issue when multiple plugins are being
executed on these 2 events. If a plugin really needs
to be executed after the page cache plugin it can still
use \Joomla\Event\Priority::MIN for its own event handler.

* Inject factories into the plugin

* Refactor the page cache plugin

* Don't try to fix stupid

* Make the plugin class final

* Using low priority

This solves the issue when multiple plugins are being
executed on these 2 events. If a plugin really needs
to be executed after the page cache plugin it can still
use \Joomla\Event\Priority::MIN for its own event handler.

* Inject factories into the plugin

* Remove version attribute from the XML manifest

* Remove useless line

* Make protected members private

I left the $app protected to prevent static code
analysers from concluding that the property must be
removed because it's never set. This wrong detection is
the consequence of the way CMSPlugin detects and sets
this property through Reflection.

* Inject the profiler

* Fix docblock

* Inject frontend router object

* Update plugins/system/cache/services/provider.php

* Update plugins/system/cache/services/provider.php

* Update plugins/system/cache/services/provider.php

* Update plugins/system/cache/services/provider.php