[4.x] Improve IE detection warning
Currently if you try to login to the admin using internet explorer you will see a warning message.

This message is hidden from all other browsers with css.

The recent changes with child templates etc resulted in some users loading there site without any css. As a result they saw the warning which of course was not helpful as they were not using IE.

This PR extends the test for ie by using the api to absolutely ensure the message will only be displayed in the targeted browsers.

Testing is quite easy
1. Using IE try and log in to the admin and you will see the warning message both with and without this pr
2. With any other browser view the source before this PR you will see the following alert.
3. With any other browser view the source after this PR and there is no alert present