[4.1] Dont inject webauthn info in the profile
This information should not be displayed in the user profile. It is only relevant to display it in the edit form.

This PR adds a check for that and also removes code which will no longer be needed.

To test create a front end menu item to user profile and set to registered.

Log in to the site and select user profile

Before PR this is displayed

After PR this is not displayed

Now click on the edit profile button and make sure that the webauthn fields are displayed there