[4.1] accessibility plugin update and fixes
#36948 changed the icons in the accessibility from the google material font (because :privacy :facepalm) to use emojis.

At the time this also resolved a bug in the upstream code that prevented the material font working on osx

Upstream has now resolved the osx problem.

This PR updates to the latest release of the script and introduces the ability to use either emoji or material icons. The default is emoji so there is no visible change etc on upgrade and no need for an update sql