[4.3] Input field direction
If a document is in LTR then all input fields are LTR and vice versa for RTL.

However there are several types of data that should always be LTR such as urls, email addresses and telephone numbers (https://rtlstyling.com/posts/rtl-styling#form-inputs).

For a field type of url we can add this to the layout.

However we usually use the field type of text for urls eg the custom url for joomla updates.

Therefore as well as adding code to the layout of the url and tel fields this pr also adds the ability to add dir=ltr to the xml definition of a field