Template Override Actions
There was a button to remove a record but not to mark it as checked or unchecked. You had to do each one individually by clicking on the icon. This PR lets you select as many records as you want and then use the button to change all the selected records. This is based on user feedback who complained that there was no way to bulk mark records.