Supported font formats mismatch
The default values in the xml include woff2
The default value in the db does not include woff2

The default values on a new install should match those in the component xml and this PR corrects that.

We can't fix it on update as that would be changing user data.

To test either code review or do a clean install from the provided build