Install Selected Languages
During the installation all the button describe what will happen when you click on them except for the next button when installing additional languages. "Next" suggests that there is something else to do before the languages are installed but there isn't.

This pr changes the string to a new string INSTL_LANGUAGES_SELECTED