Batch process on users
The batch process on users lets you remove a user from a group. But it doesn't check to make sure that the user is still in at least one group.

## To test
### Test 1
1. Create a new user and put them in the registered usergroup
2. Using the batch process try to remove them from the registered usergroup
#### Before
Success message and user now not in any group
#### After
Batch process failed with following error: A user must belong to at least one group.

### Test 2
1. Create a new user and put them in the registered AND author usergroup
2. Using the batch process try to remove them from the editor usergroup
#### Before
Success message and user still in the registered and author usergroups
#### After
Batch process failed with following error: The selected user(s) are not in the selected group.

### Test 3
1. Create a new user and put them in the registered AND author usergroup
2. Using the batch process try to remove them from the registered usergroup
#### Before
Success message and user only in the author usergroup
#### After
Success message and user only in the author usergroup

#### Comments
This type of code is really not my strong point - it may need optimizing etc