Pagination Only child
Pull request for #38512

When there is only a next button it is no longer on the right as intended. this was due to a change in the markup.

As this a change in the scss you will need to either `npm run build:css` or use one of the prebuilt packages