Language override both
When you create a language override there is an option to create the override in both the admin and site language. However this is only available if you have selected as admin language.

I can see no logical reason for this restriction today. Its very frustrating to have to create an override twice just because you selected site as the client first.

This PR removes the restriction

To test create a new override by selecting the language and site
Create an override and check the box for both locations
In the list of overrides you should now see it listed in both locations