Admin Module Cpanel Positions
Pr for #30286

In the installation data some of the modules did not have an ordering set. This created a problem as seen in #30286 when attempting to change the order of modules in the cpanel position from the module setting. (it only worked with drag and drop).

By ensuring that all the modules have a position this problem is resolved. However as this is user data we can only do this for new installs.

(Note once you do a drag and drop then the positions are corrected on existing installs)

To test.
I suggest that you first verify the problem on an existing CLEAN install.
Then create a NEW install from the prebuilt package and then try to repeat the error. Its a success if you can no longer repeat the error