Use ellipsis - part 1
Following the style guide and accessibility recommendations this PR changes the use of ... (dot dot dot) to the html entity for an ellipsis.

For example some screenreaders will announce the three dots eg read more dot dot dot but with an ellipsis they pause and do not say dot dot dot

This PR is only for the language strings. A further PR will follow to address truncation