Template Override Actions (#38447)
* Template Override Actions

There was a button to remove a record but not to mark it as checked or unchecked. You had to do each one individually by clicking on the icon. This PR lets you select as many records as you want and then use the button to change all the selected records. This is based on user feedback who complained that there was no way to bulk mark records.

* Update HtmlView.php

cs

* Update HtmlView.php

* Update default_updated_files.php

* Update com_templates.ini

* Update administrator/components/com_templates/src/View/Template/HtmlView.php

Co-authored-by: heelc29 <66922325+heelc29@users.noreply.github.com>