Duplicated code
This PR removes a block of duplicated code. They two blocks are slightly different as the home is in a different place. But as the first block is overridden by the second block I have deleted the first block.

There is no visible changes as a result of this PR. Either test by code review or by checking that you can still create and edit menu items exactly the same as before this PR