[4.2] codemirror full screen
In com_templates if you display the original file as well as the override then you have two instances of codemirror with the original file as readonly. Both have the message "Press xx to toggle Full Screen editing"  which not only doesnt make sense on a readonly instance of the editor and not surprisingly doesnt work.

This PR removes the message and associated options from the codemirror layout if the editor is defined as readonly in the xml