[4.2] Template Preview modal
It is possible that the template->name does not match the template->element

When that is the case then it is not possible to open the larger preview image of the template as the id of the modal is not the same as the id referenced in the button.

This PR corrects that by using template->element in both places.

To test
The easiest way will be to create a child of cassiopeia and then edit the database.

in #__extensions find the record for the child template and change the name only.

Now in the admin template manager you will find that you get a js error when you try to open the larger preview image.

Apply this PR and you will find that the preview opens correctly