[4.2] Media-infobar info z-index
There is a z-index issue with the media manager when you enable the infobar.

To test you will need to have quite a lot of images in the folder. I just copied all the cassiopeia sample images a few times.

## Before
When scrolling in image mode the sticky toolbar is cropped partially - see video
When scrolling in list mode the sticky toolbar is cropped a bit more - see video

## After
The entire toolbar is sticky and visible see video