[4,3] joomla alert icons
It always bugs me that the icon in the alert is off-center. (and someone else because there was an issue in the past but I cant find it now).

This simple PR sets the correct line height which re-centers the icon which I assume was the original intent based on the original css and the bootstrap alert it is inspired by.

To test you will need to either use a prebuilt package or apply the pr and `npm run build:css`

## Before

## After