[4.2] typos/grammar
I was testing a script using the joomla repo and found these typos/grammar errors.

The basic english language rule is that if you want to write "a something" and the "something" begins with a vowel then it changes to "an"

This also applies even if the "something" _sounds_ like it begins with a vowel.

As these are all corrections in comments testing is just by review