[4.2] partial revert of #38638
**I was wrong and should have listened more to what @hleithner was saying.** He was obviously correct that you can't use &hellip, in a language string in the cli.

I still think the lack of readability of using utf characters in a search string is an issue so I have reverted the use to ...

I have also reviewed all the other changes that I made in that PR and instead of reverting I am removing the &hellip; as it really served no purpose

As far as I can see now the remaining uses from that PR all relate to Read more where I still believe it is best to use &hellip but I would be happy to revert those as well.

Sorry for causing this issue.

I am not sure if we are in a freeze for strings in 4.2 so if required I can rebase to 4.3