[4.3] Cassiopeia Fonts
Currently cassiopeia offers three font configuration settings all with their limitations

1. none - boring generic system font
2. Roboto Local - slow
3. Google Fonts - privacy

Using a system font stack is always the best. It is fast, there is nothing to download, there is no layout shift (google pagespeed hates that).

The problem with a system font stack is that every system has different fonts available so you need to be careful in defining it. Thats why the current system font stack from Bootstrap that is used in Cassiopeia is so boring and safe and why people choose google fonts etc.

This PR adds 14 modern system font stacks from @danklammer

> System font stack CSS organized by typeface classification for every modern OS
The fastest fonts available. No downloading, no layout shifts, no flashes — just instant renders.

You can see them in action at https://modernfontstacks.com/ and helpfully see which font in each stack your own system will use.

This was developed for my own sites and I am offering it here as I am certain that it will be useful to all. Especially to those people who don't realise they can have great font choices with Joomla without impacting privacy or performance.

The base branch for this PR is 4.3 simply because that's what is developed on. I leave it to maintainers to decide if and when it can be merged.

It is fully backwards compatible
I created language strings for each font stack but I suspect that most translation teams will chose not to translate them.