Category Item count takes publish dates into account
Item count on a category takes the publish_up and publish_down dates into account.

## Actual result BEFORE applying this Pull Request
When setting a publish_down date in the past to unpublish an article the item count on a category will still count this article as published. And vice versa with  a publish_up date in the future.
The Categories view shows a different amount of articles than actually visible on the Category Blog view

## Expected result AFTER applying this Pull Request
Categories view will show the exact number of articles published in a category.