[4.3] Deprecate debug plugin strings (#40469)
When the debug plugin was refactored for j4 using phpdugbar a large number of strings were remaining that are simply not used. Basically only the strings in the plugin settings are used. The plugin itself is only in english and does not use translatabvle strings.

This PR deprecates all 45 unused strings so that they can be removed in 5.0 and TT have less work to do