Update to all files to get almost clean PR. Do not Review this one, wait for next please.
Some files are showing in git some are not so this is all 10 files updated to the correct code. The next release will remove the extra spaces in this release which I've added to make sure all 10 files upload. Please do not review this PR but wait for the next one in about 10 mins.