[4.3] Inclusive Naming master (comment)
We should be using inclusive naming throughout the cms not just in the language strings.

In this specific case it is not being changed as sadly php still uses the term "master value" and here it is used to produce a downloadable text file of phpinfo().

The comment is simply to explain why it has not been changed for future reference

### Rationale
https://inclusivenaming.org/word-lists/tier-1/_master/

### Test
Code review Only

### Documentation
New page will be added to manual.joomla.org shortly

### Note
This is part of a series of pull requests that are submitted separately to make reviewing easier