Completely reworked
The user plugin has been removed and introduced as a trait for the plugins in the "beforedeleteuser" plugin group. This gives the developers of third party extensions more flexibility to intervene in the exchange process themselves.

All core extensions are now combined into one "beforedelteuser" plugin.
"Modified By" is now also taken into consideration.

The global configuration of the settings has been removed, instead a popup is opened before deletion and the required settings are requested.

All adjustments could only be realized because bembelimen gave me some good tips.

Surely there is still room for improvement, but I hope you are a bit proud of me ;-)