[5.0] Missing  'scope' attribute (#41753)
Associating <table> headers, i.e. <th> elements, with their <td> cells enables screen readers to announce the header prior to the data. This considerably increases the accessibility of tables to visually impaired users.

The scope=col is missing from the folder column of the Extensions: Update view

If you dont have any extensions listed as needing an update then you can install the German lang pack and this will trigger an entry

Code review or view source will show that the "folder" column is missing a scope before this PR