[5.0] Select Schedule Task css
The select task view is the same as the select module view except it uses its own css file with invalid css.

This PR changes that so that it uses an scss file in exactly the same way as com_modules. This removes the invalid css and makes it more maintainable.