[5.0] correct usage of todo
Most IDE have tools to quickly find @todo in the code base. This PR corrects the usage of TODO instead of @todo. It will make it easier in future to work through these items, check that they are still valid, and resolve them.

code review only

Same pr as #36612 for newly added todo