[5] Add Tag Relationship in ContentSerializer.php
The work was already done by Joomla Team and other contributors what I did is just to use the TagApiSerializerTrait in ContentSerializer. It should be added everywhere where the tags are supported but to make it simple I added it just to ContentSerializer. Hope it helps. Tested it locally on my setup with Postman and it works.