Properly detect db engine support before converting tables to MEMORY
On MySQL Galera replication clusters MEMORY tables are not supported, but MySQL returns an error only at the first write. If the table is empty the first write never happens and no error is returned
at the ALTER TABLE instruction, but it will fail at the first write (i.e. editing an article). 
Proceed with the conversion to MEMORY only if the tables are not empty.