[5.1] Generic doc icon
Replaces the pdf icon used in media manager with a more generic doc icon as this is applied to a variety of doc types.

### To test
Use one of the prebuilt packages for this PR and check to see that the icon used in media manager for any doc type is no longer a pdf icon.

Changing the icon to a specific one for each type of doc is beyond the scope of this PR