[5.0] Fix missing 'separator' lines between sub menus in Admin Components
Adds separators as submenu's items for Components

Pull Request for Issue #41123 . And reflect change as in #42588 for 5.0

Summary of Changes

Alters the 'type' attribute for sub-menus when installing a Component to allow for separator lines to be added to the sub menu of a Component. For use when you need to break up or group sub-menu items of a Component.

In the manifest .xml file for a components you can indicate the placement of the separator by adding the element <menu type=separator>--<menu> between a set of elements.
---

The processing to read and render the separator already exists in the Joomla core, all that was missing was the ability to specify the 'type' of separator when installing a component. The rendering of the separator will either use the characters defined between the
and elements or the characters are stripped and the separator is rendered as a line with CSS.

Testing Instructions

Add the <menu type=separator>--<menu> element to the manifest .xml file of a component and Install the component in an instance of Joomla.
Actual result BEFORE applying this Pull Request

Nothing happened. 'type=separator' is ignored and replaced with 'component'
Expected result AFTER applying this Pull Request

Expand your Component name under the Admin area Component Menu on the left and you should see the separator displayed as a sub menu item.

In the #__menu table the entry for your separator will have a field of type with the value 'separator' and the alias will be the current date
Link to documentations

Please select:

Documentation link for docs.joomla.org: https://docs.joomla.org/Manifest_files

No documentation changes for docs.joomla.org needed

Pull Request link for manual.joomla.org:
No documentation changes for manual.joomla.org needed