[5.1] add focus to treeselect [a11y]
In the module manager when you open the tree to assign a module to specific menu items it was not possible to navigate to the chevron with a keyboard to open and close the tree.

This pr ensures that you can tab to the chevron and press enter to expand or collapse the tree.

Don't moan at me for bad js as the original js was written by the person who has asked for this bug fix